backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0
google-generativeai>=0.8.0
openai>=1.57.0
anthropic>=0.39.0
pinecone-client>=5.0.0
vaderSentiment>=3.3.2
python-multipart>=0.0.18
websockets>=14.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
